--- OpenCV_LiveCameraRecognition/build.properties
@@ -1,6 +1,0 @@
-# Project specific information
-package=org.usfirst.frc.team1732.robot
-robot.class=${package}.Robot
-simulation.world.file=/usr/share/frcsim/worlds/GearsBotDemo.world
-#Uncomment and point at user libraries to include them in the build. Do not put libraries in the \wpilib\java folder, this folder is completely overwritten on plugin update.
-#userLibs=${user.home}/wpilib/user/lib
